Move architecture tests to separate package
