+++ src/test/java/io/jenkins/plugins/archunit/PackageArchitectureTest.java
@@ -1,11 +1,11 @@
-package io.jenkins.plugins.util;
-
-import java.net.URL;
+package io.jenkins.plugins.archunit;
 
 import com.tngtech.archunit.core.importer.ImportOption.DoNotIncludeTests;
 import com.tngtech.archunit.junit.AnalyzeClasses;
 import com.tngtech.archunit.junit.ArchTest;
 import com.tngtech.archunit.lang.ArchRule;
+
+import java.net.URL;
 
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.*;
 import static com.tngtech.archunit.library.plantuml.rules.PlantUmlArchCondition.Configuration.*;
